refactor: restructure Video module
Video now has two subpackages: api & impl. Code is currently in impl.
