+++ feature/video/impl/src/main/AndroidManifest.xml
