+++ feature/video/impl/src/main/kotlin/com/potatosheep/kite/feature/video/impl/VideoScreen.kt
@@ -1,4 +1,4 @@
-package com.potatosheep.kite.feature.video
+package com.potatosheep.kite.feature.video.impl
 
 import android.content.ContentResolver
 import android.content.Context
@@ -58,7 +58,6 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.LifecycleResumeEffect
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.media3.common.Player
-import androidx.media3.common.util.Log
 import androidx.media3.common.util.UnstableApi
 import androidx.media3.exoplayer.ExoPlayer
 import androidx.media3.ui.PlayerView
