+++ feature/video/impl/src/main/res/mipmap-anydpi/ic_launcher.xml
