+++ feature/video/impl/src/main/res/values/strings.xml
