+++ settings.gradle.kts
@@ -48,7 +48,8 @@ include(":feature:user:api")
 include(":feature:search:impl")
 include(":feature:search:api")
 include(":core:markdown")
-include(":feature:video")
+include(":feature:video:impl")
+include(":feature:video:api")
 include(":core:domain")
 include(":feature:settings:impl")
 include(":feature:settings:api")
